Fix Winner bug
This was a fun one. Even when all my integration tests would pass, I would still see
"sh: 1: Winner not found" in the output. A Google search suggested that the text
'Winner is "Star Wars"!' (from L#47 of Results_spec.rb) was being piped to shell
and a test run with a simple modification of the text proved this.

After spending some time debating how to track down this bug if it was in Capybara or
Selenium's source code, I realized that I had successfully used "expect(page).to have_text"
in another example in the same spec file and that the actual culprit was the backticks I
had mistakenly enclosed the text in (one of the casualties of switching between Node & Rails).